+++ labelme/__init__.py
@@ -12,5 +12,3 @@ __version__ = importlib.metadata.version("labelme")
 # XXX: has to be imported before PySide6 to load dlls in order on Windows
 # https://github.com/wkentaro/labelme/issues/1564
 import onnxruntime
-
-from labelme import utils